refactor(Loader): migrate from JSX to TypeScript and optimize styles
Migrate the Loader component from JSX to TypeScript for improved type safety and maintainability. Optimize CSS styles by using shorthand properties and adding comments for better readability. Add a new SVG asset for visual enhancements.
