+++ src/layouts/Layout.astro
@@ -1,5 +1,5 @@
 ---
-import Loader from "@/components/Loader.jsx";
+import Loader from "@/components/Loader.tsx";
 import { ViewTransitions } from "astro:transitions";
 // Importa las fuentes que necesites
 import "@fontsource-variable/raleway";
